Fix a lot of bugs in the PC editor - Alchemy dialog had black text on dark background - Buttons didn't depress when clicked - Edit XP button just said "Edit" - Two menuitems in Special Edit menu had been left with default names - Callbacks in Special Edit and Free Extras menu weren't correctly assigned - Edit XP dialog put the XP to next level in the wrong place - Fixed alchemy dialog LED naming
Supporting changes in the dialog engine:
- Getting or setting control text as a number now uses the long long type to allow maximal precision
- When adding a label to a button, the dialog's default text colour is used instead of the button's text colour if the dialog's background is dark

+++ osx/dialogxml/control.cpp
@@ -316,15 +316,15 @@ eControlType cControl::getType(){
 	return type;
 }
 
-void cControl::setTextToNum(long what){
+void cControl::setTextToNum(long long what){
 	std::ostringstream sout;
 	sout << what;
 	setText(sout.str());
 }
 
-long cControl::getTextAsNum(){
+long long cControl::getTextAsNum(){
 	std::istringstream sin(getText());
-	long n;
+	long long n;
 	sin >> n;
 	return n;
 }
